+++ lib/presentation/widgets/attachment/on_attachment.dart
@@ -6,7 +6,7 @@ import 'dart:convert';
 import 'package:accredit/core/utils/my_encryption.dart';
 import 'package:accredit/core/utils/my_logs.dart';
 import 'package:accredit/dal/local/local_source_adapter.dart';
-import 'package:accredit/domain/models/mode_buckets.dart';
+
 import 'package:accredit/domain/models/source_item.dart';
 import 'package:accredit/domain/services/card_items_manager.dart';
 import 'package:flutter/material.dart';